chore: move hook
